chore: refactor article model test
